requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
greenlet==2.0.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pydantic==1.10.9
python-dotenv==1.0.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
typing-extensions==4.6.3
Werkzeug==2.3.6
zipp==3.15.0
